feat: refactor route imports to simplify path structure for analytics modules
